feat:로고/슬로건 컴포넌트 분리
